# Support async IO for MultiScan