--- a/db/version_set.cc
+++ b/db/version_set.cc
@@ -1147,6 +1147,7 @@ class LevelIterator final : public InternalIterator {
     // Propagate io colaescing threshold
     for (auto& file_to_arg : *file_to_scan_opts_) {
       file_to_arg.second.io_coalesce_threshold = so->io_coalesce_threshold;
+      file_to_arg.second.use_async_io = so->use_async_io;
     }
   }