--- a/db_stress_tool/db_stress_gflags.cc
+++ b/db_stress_tool/db_stress_gflags.cc
@@ -1531,4 +1531,7 @@ DEFINE_bool(
 DEFINE_bool(use_multiscan, false,
             "If set, use the batched MultiScan API for scans.");
 
+DEFINE_bool(use_async_io_multiscan, false,
+            "If set, enable async_io for MultiScan operations.");
+
 #endif  // GFLAGS