--- a/table/block_based/block_based_table_iterator.h
+++ b/table/block_based/block_based_table_iterator.h
@@ -398,12 +398,7 @@ class BlockBasedTableIterator : public InternalIteratorBase<Slice> {
     bool finished{false};
 
     AsyncReadState() = default;
-
-    // Explicitly default move operations since we have a user-defined
-    // destructor.
-    AsyncReadState(AsyncReadState&&) = default;
-    AsyncReadState& operator=(AsyncReadState&&) = default;
-
+    DECLARE_DEFAULT_MOVES(AsyncReadState);
     // Delete copy operations
     AsyncReadState(const AsyncReadState&) = delete;
     AsyncReadState& operator=(const AsyncReadState&) = delete;